apps/ai-chat-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
python-multipart==0.0.6
openai==1.3.0
anthropic==0.7.0
boto3==1.34.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.25.0
litellm==1.17.0
langfuse==3.5.2
redis==5.0.1
pillow==10.1.0
